server/requirements.txt
fastapi
uvicorn
pydantic
pymongo[srv]
python-dotenv
scrapy
boto3
openai==1.16.2